# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Referaa - Refer, mentor, and connect platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all public job listings
      parameters:
        - name: company
          in: query
          schema:
            type: string
        - name: title
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: remote
          in: query
          schema:
            type: string
        - name: experienceMin
          in: query
          schema:
            type: integer
        - name: salaryMin
          in: query
          schema:
            type: integer
        - name: scope
          in: query
          description: community excludes your own jobs; all shows every active job
          schema:
            type: string
            enum: [community, all]
            default: community
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job listing
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job listing
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{jobId}/referrals:
    get:
      operationId: listJobReferrals
      tags: [jobs]
      summary: List referral requests for a job (poster only)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Referrals for this job
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferralSummary"
        "404":
          description: Job not found
  /jobs/{jobId}/save:
    post:
      operationId: saveJob
      tags: [jobs]
      summary: Save a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedStatus"
    delete:
      operationId: unsaveJob
      tags: [jobs]
      summary: Unsave a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsaved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedStatus"

  # Referrals
...
  /messages/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        headline:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        currentRole:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        resumeScore:
          type: ["integer", "null"]
        successRate:
          type: ["integer", "null"]
        totalPoints:
          type: integer
        isWorkingProfessional:
          type: boolean
        isConsultant:
          type: boolean
        memberType:
          type: string
          enum: [student, alumni]
        workExperiences:
          type: array
          items:
            type: object
            properties:
              company:
                type: string
              role:
                type: ["string", "null"]
              fromYear:
                type: ["string", "null"]
              toYear:
                type: ["string", "null"]
              description:
                type: ["string", "null"]
            required:
              - company
        projects:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: ["string", "null"]
              technologies:
                type: ["string", "null"]
            required:
              - name
        education:
          type: array
          items:
            type: object
            properties:
              level:
                type: string
                enum: [UG, PG, PhD]
              institution:
                type: string
              stream:
                type: ["string", "null"]
              batchYear:
                type: ["integer", "null"]
            required:
              - level
              - institution
        researchPapers:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              publication:
                type: ["string", "null"]
              year:
                type: ["integer", "null"]
              link:
                type: ["string", "null"]
            required:
              - title
        certifications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              issuer:
                type: ["string", "null"]
              year:
                type: ["integer", "null"]
            required:
              - name
        isProfileComplete:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - fullName
        - email
        - skills
        - totalPoints
        - isWorkingProfessional
        - isConsultant
        - memberType
        - workExperiences
        - projects
        - education
        - researchPapers
        - certifications
        - isProfileComplete
        - createdAt
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogoUrl:
          type: ["string", "null"]
        location:
          type: string
        isRemote:
          type: boolean
        isHybrid:
          type: boolean
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceMin:
          type: ["integer", "null"]
        experienceMax:
          type: ["integer", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        salaryDisclosed:
          type: boolean
        rewardPoints:
          type: integer
        status:
          type: string
          enum: [active, closed, paused]
        posterId:
          type: integer
        poster:
          $ref: "#/components/schemas/UserProfile"
        isSaved:
          type: boolean
        referralCount:
          type: integer
        isOwnJob:
          type: boolean
        myReferralStatus:
          type: ["string", "null"]
          enum: [pending, accepted, rejected, referred, interviewing, hired, rejected_after_interview, null]
        myReferral:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            status:
              type: string
            note:
              type: ["string", "null"]
            totalPointsCredited:
              type: integer
            totalPointsDeducted:
              type: integer
            createdAt:
              type: string
            updatedAt:
              type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - company
        - location
        - isRemote
        - isHybrid
        - description
        - skills
        - salaryDisclosed
        - rewardPoints
        - status
        - posterId
        - poster
        - isSaved
        - referralCount
        - isOwnJob
        - createdAt
    JobInput:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        companyLogoUrl:
          type: string
        location:
          type: string
        isRemote:
          type: boolean
        isHybrid:
          type: boolean
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceMin:
          type: integer
        experienceMax:
          type: integer
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        salaryDisclosed:
          type: boolean
        rewardPoints:
          type: integer
      required:
        - title
        - company
        - location
        - description
        - skills
        - salaryDisclosed
        - rewardPoints
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        isRemote:
          type: boolean
        isHybrid:
          type: boolean
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceMin:
          type: integer
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        salaryDisclosed:
          type: boolean
        rewardPoints:
          type: integer
        status:
          type: string
          enum: [active, closed, paused]
    SavedStatus:
      type: object
      properties:
        saved:
          type: boolean
      required:
        - saved
...
    ReferralSummary:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        requesterId:
          type: integer
        requester:
          $ref: "#/components/schemas/UserProfile"
        referrerId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected, referred, interviewing, hired, rejected_after_interview]
        note:
          type: ["string", "null"]
        rewardStagesApplied:
          type: array
          items:
            type: string
        totalPointsDeducted:
          type: integer
        totalPointsCredited:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - jobId
        - requesterId
        - requester
        - referrerId
        - status
        - rewardStagesApplied
        - totalPointsDeducted
        - totalPointsCredited
        - createdAt
        - updatedAt
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: string
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/UserProfile"
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - senderId
        - sender
        - content
        - createdAt
    MessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...